// TRC-Beneficios-Front/src/lib/interfaces/IBeneficiario.ts
export interface IBeneficiario {
    cpf: string
    nome?: string
    data_nascimento?: string
    genero?: string
    rua?: string
    telefone?: string
    celular?: string
    rg?: string
    data_obito?: string
    email?: string
    representante_legal?: string
    c_documentos?: string
    cep?: string
    bairro?: string
    cidade?: string
    num_complemento?: string
    num?: string
    complemento?: string
    deficiencia?: string
}